+++ src/ts/Managers/OperativeEventManager.ts
@@ -247,6 +247,9 @@ export class OperativeEventManager {
             infoJson.ts = (new Date()).toISOString();
             infoJson.event_type = eventType;
             infoJson.sourceGameId = this._clientInfo.getGameId();
+            if(campaign instanceof XPromoCampaign) {
+                infoJson.targetGameId = campaign.getGameId().toString();
+            }
 
             HttpKafka.sendEvent(kafkaType, infoJson);
         };