+++ src/ts/Test/Unit/Managers/CampaignRefreshManagerTests.ts
@@ -497,7 +497,7 @@ describe('CampaignRefreshManager', () => {
             return campaignRefreshManager.refresh().then(() => {
                 diagnosticsStub.restore();
                 assert.equal(receivedErrorType , 'auction_request_failed', 'Incorrect error type');
-                assert.equal(receivedError.error.message ,'test error', 'Incorrect error message');
+                assert.equal(receivedError.error.message , 'test error', 'Incorrect error message');
             });
         });
 
@@ -517,7 +517,7 @@ describe('CampaignRefreshManager', () => {
             return campaignRefreshManager.refresh().then(() => {
                 diagnosticsStub.restore();
                 assert.equal(receivedErrorType , 'auction_request_failed', 'Incorrect error type');
-                assert.equal(receivedError.error.message ,'test error', 'Incorrect error message');
+                assert.equal(receivedError.error.message , 'test error', 'Incorrect error message');
             });
         });
 
@@ -544,7 +544,7 @@ describe('CampaignRefreshManager', () => {
             return campaignRefreshManager.refresh().then(() => {
                 diagnosticsStub.restore();
                 assert.equal(receivedErrorType , 'auction_request_failed', 'Incorrect error type');
-                assert.equal(receivedError.error.message ,'Unsupported content-type: wrong/contentType', 'Incorrect error message');
+                assert.equal(receivedError.error.message , 'Unsupported content-type: wrong/contentType', 'Incorrect error message');
             });
         });
 
@@ -570,8 +570,8 @@ describe('CampaignRefreshManager', () => {
 
             return campaignRefreshManager.refresh().then(() => {
                 diagnosticsStub.restore();
-                assert.equal(receivedErrorType , 'auction_request_failed', 'Incorrect error type');
-                assert.equal(receivedError.error.message ,'model: AuctionResponse key: contentType with value: 1: integer is not in: string', 'Incorrect error message');
+                assert.equal(receivedErrorType, 'auction_request_failed', 'Incorrect error type');
+                assert.equal(receivedError.error.message, 'model: AuctionResponse key: contentType with value: 1: integer is not in: string', 'Incorrect error message');
             });
         });
     });